Module actix_web_location::providers

Providers are actix-web-location’s abstraction to allow multiple ways of determining location.

Structs

FallbackProvider

A “dummy” provider that returns None for all fields.

MaxMindProvider

A provider that uses a MaxMind GeoIP database to derive location from a the IP a request was sent from.

Traits

Provider

An object that can be queried to convert HttpRequest into locations.